Refactor widget to other file
