+++ app/(detail)/[patientId]/dashboards/page.tsx
@@ -1,8 +1,6 @@
 'use client';
-import AddButton from '@/components/Common/AddButton';
-import FollowUpsDashboard from '@/components/Dashboards/FollowUps/FollowUpsDashboard';
-import FollowUpForm from '@/components/Forms/FollowUpForm';
-import { Box, Sheet, Typography } from '@mui/joy';
+import FollowUpWidget from '@/components/Dashboards/FollowUps/FollowUpWidget';
+import { Sheet } from '@mui/joy';
 
 interface DashboardsPageProps {
   params: {
@@ -22,37 +20,10 @@ export default function DashboardsPage({ params }: DashboardsPageProps) {
         justifyContent: 'space-between',
       }}
     >
-      <FollowUpWidget width={width} patientId={patientId}/>
-      <FollowUpWidget width={width} patientId={patientId}/>
-      <FollowUpWidget width={width} patientId={patientId}/>
-      <FollowUpWidget width={width} patientId={patientId}/>
-      
+      <FollowUpWidget width={width} patientId={patientId} />
+      <FollowUpWidget width={width} patientId={patientId} />
+      <FollowUpWidget width={width} patientId={patientId} />
+      <FollowUpWidget width={width} patientId={patientId} />
     </Sheet>
-  );
-}
-interface FollowUpWidgetProps {
-  width: number;
-  patientId: string;
-}
-function FollowUpWidget({ width, patientId }: FollowUpWidgetProps) {
-  return (
-    <Box
-      sx={{
-        width: `${width}%`,
-        display: 'flex',
-        flexDirection: 'column',
-        justifyContent: 'flex-end',
-        rowGap: 2,
-      }}
-    >
-      <Box sx={{ display: 'flex', justifyContent: 'space-between' }}>
-        <Typography level="h3">Seguimientos</Typography>
-        <AddButton
-          text="Crear Seguimiento"
-          form={<FollowUpForm patientId={patientId} />}
-        />
-      </Box>
-      <FollowUpsDashboard patientId={patientId} />
-    </Box>
   );
 }
